refactor: migrate announcements module to main GraphQL schema
The shop.announcements API was backported to 3.23, so it now exists in
the real main schema. Migrate the module off the staging track:

- queries.staging.ts -> queries.ts, fragment likewise
- @dashboard/graphql/staging imports -> @dashboard/graphql
- drop the skip: !isStagingSchema() guard in useAnnouncements
- revert staging schema/generated files to baseline; main generated
  files now carry the announcements query/hook/types
